moves dht RoutingTable to testrouting
