+++ pkg/dht/dht.go
@@ -5,11 +5,9 @@ package dht
 
 import (
 	"context"
-	"time"
 
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/storage"
 )
 
 // DHT is the interface for the DHT in the Storj network
@@ -19,23 +17,6 @@ type DHT interface {
 	Ping(ctx context.Context, node pb.Node) (pb.Node, error)
 	FindNode(ctx context.Context, ID storj.NodeID) (pb.Node, error)
 	Seen() []*pb.Node
-}
-
-// RoutingTable contains information on nodes we have locally
-type RoutingTable interface {
-	// local params
-	Local() pb.Node
-	K() int
-	CacheSize() int
-	GetBucketIds() (storage.Keys, error)
-	FindNear(id storj.NodeID, limit int, restrictions ...pb.Restriction) ([]*pb.Node, error)
-	ConnectionSuccess(node *pb.Node) error
-	ConnectionFailed(node *pb.Node) error
-	// these are for refreshing
-	SetBucketTimestamp(id []byte, now time.Time) error
-	GetBucketTimestamp(id []byte) (time.Time, error)
-
-	Close() error
 }
 
 // Bucket is a set of methods to act on kademlia k buckets
