+++ pkg/kademlia/routing_integration_test.go
@@ -7,17 +7,15 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
 	"storj.io/storj/internal/testcontext"
-	"storj.io/storj/pkg/dht"
 	"storj.io/storj/pkg/kademlia/testrouting"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
 )
 
-type routingCtor func(storj.NodeID, int, int, int) dht.RoutingTable
+type routingCtor func(storj.NodeID, int, int, int) testrouting.RoutingTable
 
-func newRouting(self storj.NodeID, bucketSize, cacheSize, allowedFailures int) dht.RoutingTable {
+func newRouting(self storj.NodeID, bucketSize, cacheSize, allowedFailures int) testrouting.RoutingTable {
 	if allowedFailures != 0 {
 		panic("failure counting currently unsupported")
 	}
@@ -27,7 +25,7 @@ func newRouting(self storj.NodeID, bucketSize, cacheSize, allowedFailures int) d
 	})
 }
 
-func newTestRouting(self storj.NodeID, bucketSize, cacheSize, allowedFailures int) dht.RoutingTable {
+func newTestRouting(self storj.NodeID, bucketSize, cacheSize, allowedFailures int) testrouting.RoutingTable {
 	return testrouting.New(self, bucketSize, cacheSize, allowedFailures)
 }
